requirements.txt
flask
flask-cors
flask-socketio
pycloudflared
python-dotenv
openai
edge-tts
ollama
SpeechRecognition
pydub
Pillow
pytesseract
requests
wakeonlan
samsungtvws
pyautogui
duckduckgo-search
beautifulsoup4
qrcode